replace doubles with floats

+++ verb.c
@@ -44,7 +44,7 @@ My github:     https://github.com/el-visio
 #define MAX_PREDELAY 4800  // 100ms for 48k samplerate
 
 /* Clamp value between min and max */
-double clamp(double x, double min, double max)
+float clamp(float x, float min, float max)
 {
   if (x < min)
     return min;
@@ -82,12 +82,12 @@ void DelayBuffer_init(DelayBuffer* db, uint16_t delay)
   bufferSize = 1 << numBits;
 
   // Allocate buffer
-  db->buffer = malloc(bufferSize * sizeof(double));
+  db->buffer = malloc(bufferSize * sizeof(float));
   if (!db->buffer)
     return;
 
   // Clear buffer
-  memset(db->buffer, 0, bufferSize * sizeof(double));
+  memset(db->buffer, 0, bufferSize * sizeof(float));
 
   // Create bitmask for fast wrapping of the circular buffer
   db->mask = bufferSize - 1;
@@ -102,79 +102,79 @@ void DelayBuffer_delete(DelayBuffer* db)
 }
 
 /* Write input value into buffer, read delayed output */
-double DelayBuffer_process(DelayBuffer* db, uint16_t t, double in)
+float DelayBuffer_process(DelayBuffer* db, uint16_t t, float in)
 {
   db->buffer[t & db->mask] = in;
   return db->buffer[(t + db->readOffset[TAP_MAIN]) & db->mask];
 }
 
 /* Write value into delay buffer */
-void DelayBuffer_write(DelayBuffer* db, uint16_t t, double in) 
+void DelayBuffer_write(DelayBuffer* db, uint16_t t, float in) 
 {
   db->buffer[t & db->mask] = in;
 }
 
 /* Read delayed output value */
-double DelayBuffer_read(DelayBuffer* db, uint16_t tapId, uint16_t t) 
+float DelayBuffer_read(DelayBuffer* db, uint16_t tapId, uint16_t t) 
 {
   return db->buffer[(t + db->readOffset[tapId]) & db->mask];
 }
 
 /* Apply all-pass filter */
-double AllPassFilter_process(DelayBuffer* db, uint16_t t, double gain, double in)
+float AllPassFilter_process(DelayBuffer* db, uint16_t t, float gain, float in)
 {
-  double delayed = DelayBuffer_read(db, TAP_MAIN, t);
+  float delayed = DelayBuffer_read(db, TAP_MAIN, t);
   in += delayed * -gain;
   DelayBuffer_write(db, t, in);
   return delayed + in * gain;
 }
 
 /* Apply Low pass filter */
-double LowPassFilter_process(double* out, double freq, double in)
+float LowPassFilter_process(float* out, float freq, float in)
 {
   *out += (in - *out) * freq;
   return *out;
 }
 
 /* Set pre-delay length (relative to MAX_PREDELAY) */
-void DattorroVerb_setPreDelay(DattorroVerb* v, double value)
+void DattorroVerb_setPreDelay(DattorroVerb* v, float value)
 {
   DelayBuffer_setDelay(&v->preDelay, TAP_MAIN, value * MAX_PREDELAY);
 }
 
 /* Set pre-filter amount */
-void DattorroVerb_setPreFilter(struct sDattorroVerb* v, double value)
+void DattorroVerb_setPreFilter(struct sDattorroVerb* v, float value)
 {
   v->preFilterAmount = value;
 }
 
 /* Set input diffusion 1 amount */
-void DattorroVerb_setInputDiffusion1(struct sDattorroVerb* v, double value)
+void DattorroVerb_setInputDiffusion1(struct sDattorroVerb* v, float value)
 {
   v->inputDiffusion1Amount = value;
 }
 
 /* Set input diffusion 2 amount */
-void DattorroVerb_setInputDiffusion2(struct sDattorroVerb* v, double value)
+void DattorroVerb_setInputDiffusion2(struct sDattorroVerb* v, float value)
 {
   v->inputDiffusion2Amount = value;
 }
 
 /* Set decay diffusion 1 amount */
-void DattorroVerb_setDecayDiffusion(struct sDattorroVerb* v, double value)
+void DattorroVerb_setDecayDiffusion(struct sDattorroVerb* v, float value)
 {
   v->decayDiffusion1Amount = value;
 }
 
 /* Set decay amount and calculate related decay diffusion 2 amount */
-void DattorroVerb_setDecay(DattorroVerb* v, double value)
+void DattorroVerb_setDecay(DattorroVerb* v, float value)
 {
   v->decayAmount = value;
   v->decayDiffusion2Amount = clamp(value + 0.15, 0.25, 0.50);
 }
 
 /* Set damping amount */
-void DattorroVerb_setDamping(struct sDattorroVerb* v, double value)
+void DattorroVerb_setDamping(struct sDattorroVerb* v, float value)
 {
   v->dampingAmount = value;
 }
@@ -268,9 +268,9 @@ void DattorroVerb_delete(DattorroVerb* v)
 // After calling this function you can
 // get wet stereo reverb signal by calling
 // DattorroVerb_getLeft and DattorroVerb_getRight 
-void DattorroVerb_process(DattorroVerb* v, double in)
+void DattorroVerb_process(DattorroVerb* v, float in)
 {
-  double x, x1;
+  float x, x1;
 
   // Modulate decayDiffusion1A & decayDiffusion1B
   if ((v->t & 0x07ff) == 0) {
@@ -313,9 +313,9 @@ void DattorroVerb_process(DattorroVerb* v, double in)
 }
 
 // Get left channel reverb
-double DattorroVerb_getLeft(DattorroVerb* v)
+float DattorroVerb_getLeft(DattorroVerb* v)
 {
-  double a;
+  float a;
   a  = DelayBuffer_read(&v->preDampingDelay[1],  TAP_OUT1, v->t);
   a += DelayBuffer_read(&v->preDampingDelay[1],  TAP_OUT2, v->t);
   a -= DelayBuffer_read(&v->decayDiffusion2[1],  TAP_OUT2, v->t);
@@ -327,9 +327,9 @@ double DattorroVerb_getLeft(DattorroVerb* v)
 }
 
 // Get right channel reverb
-double DattorroVerb_getRight(DattorroVerb* v)
+float DattorroVerb_getRight(DattorroVerb* v)
 {
-  double a;
+  float a;
   a  = DelayBuffer_read(&v->preDampingDelay[0],  TAP_OUT1, v->t);
   a += DelayBuffer_read(&v->preDampingDelay[0],  TAP_OUT2, v->t);
   a -= DelayBuffer_read(&v->decayDiffusion2[0],  TAP_OUT2, v->t);
